pyproject.toml
[tool.poetry]
name = "delty"
version = "0.1.0"
description = ""
authors = ["Dimitris Orfanoudakis <dimitrisorfanudakis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
django-crispy-forms = "^2.2"
crispy-bootstrap5 = "^2024.2"
django-debug-toolbar = "^4.4.2"
django-redis = "^5.4.0"
django-database-url = "^1.0.3"
psycopg2 = "^2.9.9"
pre-commit = "^3.7.1"
mypy = "^1.10.1"
types-requests = "^2.32.0.20240622"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "app.settings"
